BUG: Use enable_if with SFINAE to dispatch
addresses: #1003
