+++ Modules/Filtering/ImageFilterBase/include/itkCastImageFilter.h
@@ -20,6 +20,7 @@
 
 #include "itkUnaryFunctorImageFilter.h"
 #include "itkProgressReporter.h"
+#include "itkMetaProgrammingLibrary.h"
 
 #include <type_traits>
 
@@ -133,25 +134,18 @@ protected:
 
   void DynamicThreadedGenerateData(const OutputImageRegionType & outputRegionForThread) override;
 
-  template<typename TInputPixelType>
-    void DynamicThreadedGenerateDataDispatched(const OutputImageRegionType & outputRegionForThread, std::true_type isConvertible);
+  template<typename TInputPixelType,
+           typename TOutputPixelType,
+           typename std::enable_if<mpl::is_static_castable<TInputPixelType, TOutputPixelType>::value, int>::type = 0>
+    void DynamicThreadedGenerateDataDispatched(const OutputImageRegionType & outputRegionForThread);
 
-  template<typename TInputPixelType>
-    void DynamicThreadedGenerateDataDispatched(const OutputImageRegionType & outputRegionForThread, std::false_type isConvertible);
+  template<typename TInputPixelType,
+           typename TOutputPixelType,
+           typename std::enable_if<!mpl::is_static_castable<TInputPixelType, TOutputPixelType>::value, int>::type = 0>
+   void DynamicThreadedGenerateDataDispatched(const OutputImageRegionType & outputRegionForThread);
 
 private:
 
-  template <typename TFromType, typename TToType, typename TNoDeclType = TToType>
-    struct is_static_castable
-    : std::false_type
-  {};
-
-  template <typename TFromType, typename TToType>
-    struct is_static_castable<TFromType,
-                              TToType,
-                              decltype(static_cast<TToType>(std::declval<TFromType>()))>
-    : std::true_type
-  {};
 
 };
 } // end namespace itk
